* Linspace and Sequence now works for cases where start==end. This caused a lot of subtle changes, and should probably be refactored later. * Fixed a few unit tests related to the fact that zero-sized Arrays are now supported. All 93 pass on the PC.

+++ coretech/common/include/anki/common/robot/matrix.h
@@ -576,7 +576,7 @@ namespace Anki
             point1Warped[0][0] /= point1Warped[2][0];
             point1Warped[1][0] /= point1Warped[2][0];
 
-            const Type distance = sqrtf(powf(transformedPoints[iPoint].x - point1Warped[0][0], 2.0f) + powf(transformedPoints[iPoint].y - point1Warped[1][0], 2.0f));
+            const Type distance = sqrtf(powf(static_cast<f32>(transformedPoints[iPoint].x) - static_cast<f32>(point1Warped[0][0]), 2.0f) + powf(static_cast<f32>(transformedPoints[iPoint].y) - static_cast<f32>(point1Warped[1][0]), 2.0f));
 
             if(distance > MAX_SOLVE_DISTANCE) {
               AnkiWarn("EstimateHomography", "Poor solution precision");
